feat(admin): added admin dashboard for detailed view of products

+++ backend/services/productService.js
@@ -1,6 +1,8 @@
 const prisma = require("../config/prisma");
+const logger = require("../utils/logger");
 
 const createProduct = async (data, userId) => {
+  logger.info(`Product created: ${data.title} by user ${userId}`);
   return await prisma.product.create({
     data: {
       title: data.title,
@@ -90,7 +92,21 @@ const getProductById = async (id) => {
   });
 };
 
-const updateProduct = async (id, data) => {
+const updateProduct = async (id, data, user) => {
+  logger.info(`Product updated: ${id} by user ${user.id}`);
+  const product = await prisma.product.findUnique({
+    where: { id },
+  });
+  if (!product) {
+    logger.warn(`Product not found: ${id}`);
+    throw new Error("Product not found");
+  }
+  //ownership check
+  if (user.role === "SELLER" && product.user_id !== user.id) {
+    logger.warn(`Unauthorized attempt to update product: ${id}`);
+    throw new Error("Not authorized to update this product");
+  }
+
   const cleanData = {};
 
   if (data.title) cleanData.title = data.title;
@@ -105,7 +121,20 @@ const updateProduct = async (id, data) => {
   });
 };
 
-const deleteProduct = async (id) => {
+const deleteProduct = async (id, user) => {
+  logger.info(`Product deleted: ${id} by user ${user.id}`);
+  const product = await prisma.product.findUnique({
+    where: { id },
+  });
+  if (!product) {
+    logger.warn(`Product not found: ${id}`);
+    throw new Error("Product not found");
+  }
+  //ownership check
+  if (user.role === "SELLER" && product.user_id !== user.id) {
+    logger.warn(`Unauthorized attempt to delete product: ${id}`);
+    throw new Error("Not authorized to delete this product");
+  }
   return await prisma.product.delete({
     where: { id: id },
   });
